# specs/001-better-auth-authentication/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Better Auth Authentication API
  version: 1.0.0
  description: API for user authentication, authorization, and task management.

servers:
  - url: http://localhost:8000
    description: Local Development Server

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    UserRegister:
      type: object
      required:
        - email
        - username
        - password
      properties:
        email:
          type: string
          format: email
          description: Unique email address of the user.
        username:
          type: string
          description: Unique username of the user.
        password:
          type: string
          format: password
          description: User's password.
          minLength: 8 # Example minimum length

    UserLogin:
      type: object
      required:
        - identifier
        - password
      properties:
        identifier:
          type: string
          description: User's email or username.
        password:
          type: string
          format: password
          description: User's password.

    AuthToken:
      type: object
      required:
        - access_token
        - token_type
      properties:
        access_token:
          type: string
          description: JWT access token.
        token_type:
          type: string
          description: Type of the token (e.g., "bearer").

    TaskCreate:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          description: Description of the task.

    TaskUpdate:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          description: Updated description of the task.

    Task:
      type: object
      required:
        - id
        - content
        - user_id
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the task.
        content:
          type: string
          description: Description of the task.
        user_id:
          type: string
          format: uuid
          description: Unique identifier of the user who owns the task.

    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: A human-readable explanation of the error.

paths:
  /users/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '201':
          description: User successfully registered.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                    format: uuid
                  email:
                    type: string
                    format: email
                  username:
                    type: string
        '400':
          description: Invalid input or email/username already registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /users/login:
    post:
      summary: Authenticate user and get an access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: User successfully authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthToken'
        '401':
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /users/logout:
    post:
      summary: Invalidate the current user session/token
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User successfully logged out.
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /tasks:
    get:
      summary: Get all tasks for the authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: A list of tasks.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new task for the authenticated user
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /tasks/{task_id}:
    parameters:
      - in: path
        name: task_id
        schema:
          type: string
          format: uuid
        required: true
        description: Unique identifier of the task.
    get:
      summary: Get a specific task for the authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: A single task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found or not owned by user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update an existing task for the authenticated user
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found or not owned by user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a task for the authenticated user
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Task successfully deleted.
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found or not owned by user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'